Upgrade fmt to version 9.0.0
Adjust simc to not rely on implici enum to integer conversion for fmt formatting, which has been deprecated and spits out warnings.
This means adding casts to int where necessary. In the case of rapidjson error codes, a custom formatter using the rapidjson::GetParseError_En function was added.

+++ engine/action/variable.cpp
@@ -283,7 +283,7 @@ void variable_t::execute()
   if (operation != OPERATION_PRINT)
   {
     sim->print_debug("{} variable name={} op={} value={} default={} sig={}", *player, var->name_,
-      operation, var->current_value_, var->default_value_, signature_str);
+      static_cast<int>(operation), var->current_value_, var->default_value_, signature_str);
   }
 
   switch (operation)
